+++ srcs/find_in_pool.c
@@ -6,7 +6,7 @@
 /*   By: pchadeni <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/25 16:37:40 by pchadeni          #+#    #+#             */
-/*   Updated: 2019/02/25 16:37:53 by pchadeni         ###   ########.fr       */
+/*   Updated: 2019/03/07 12:30:20 by pchadeni         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,23 +15,21 @@
 t_block	*find_block_of_ptr(void *ptr, t_alloc *type)
 {
 	t_block	*result;
+	t_block	*block;
 
 	result = NULL;
 	if (!ptr)
 		return (NULL);
+	block = (t_block *)(ptr - sizeof(t_block));
 	while (*type <= LARGE_TYPE && !result)
 	{
-		result = find_last_block(g_pool[*type],
-				(t_block *)(ptr - sizeof(t_block)));
-		if (!result || (t_block *)(ptr - sizeof(t_block)) != result)
+		result = find_last_block(g_pool[*type], block);
+		if (block != result)
 			result = NULL;
 		(*type)++;
 	}
-	if (result && !result->next &&
-			(t_block *)(ptr - sizeof(t_block)) != result)
-		return (NULL);
 	(*type)--;
-	return (result);
+	return (block == result ? result : NULL);
 }
 
 t_block	*find_available_chunk(size_t size, t_alloc type, t_block **last)
